chore: remove lifecycles scripts (#2839)
* chore: remove lifecycles scripts

* chore: manually install chrome for puppeteer

* chore: typo in yaml

* chore: update test scripts

* chore: update script

* chore: update format script

* fix: add cooldown

* fix: remove github token

* chore: pr feedback

* chore: reworking app definition tests to fix tests

* chore: add gh token back into release

* chore: remove test branch from release

---------

+++ test/integration/resource-provider-integration.test.ts
@@ -23,11 +23,9 @@ describe('ResourceProvider API', () => {
   let appBundle: AppBundle
 
   beforeEach(async () => {
-    await timeoutToCalmRateLimiting()
-
     organization = await getTestOrganization()
     appDefinition = await organization.createAppDefinition({
-      name: 'Test',
+      name: 'Test AppBundle Resource Provider Creation in CMA (delete if stale)',
       src: 'http://localhost:2222',
       locations: [{ location: 'entry-sidebar' }],
     })
@@ -46,13 +44,11 @@ describe('ResourceProvider API', () => {
   })
 
   afterEach(async () => {
-    const { items: appDefinitions } = await organization.getAppDefinitions()
-
     if (appUpload) {
       await appUpload.delete()
     }
 
-    for await (const appDefinition of appDefinitions) {
+    if (appDefinition) {
       await appDefinition.delete()
     }
